automationqa-course/4.xunit/matchers/1.What is matchers/src/main/java/ru/sbtqa/tutorials/xunit/PaymentBankAccount.java
package ru.sbtqa.tutorials.xunit;

import java.math.BigDecimal;


public class PaymentBankAccount  {

    private String firstName;
    private String lastName;
    private BigDecimal balance;

    PaymentBankAccount(String firstName, String lastName, BigDecimal balance) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.balance = balance;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public BigDecimal getBalance() {
        return balance;
    }

    public void setBalance(BigDecimal balance) {
        this.balance = balance;
    }
}
